refactor(io): Refactor network types into net::io modules
